Forgot to cast to int

+++ lib/hue/scenes/scene.dart
@@ -26,7 +26,7 @@ class Scene {
       : _name = json["name"],
         _type = json["type"],
         _group = json["group"],
-        _lights = json["lights"].map((stringId) => int.parse(stringId)).toList(),
+        _lights = json["lights"].map((stringId) => int.parse(stringId)).toList().cast<int>(),
         _owner = json["owner"],
         _recycle = json["recycle"],
         _locked = json["locked"],
